Update block-begin and block-end lines in org-mode
Give both a darker font and shrink them, turn the end line’s background the same
color as the ‘org-block’ face. This makes it clear that the begin and end lines
aren’t as important as the contents and by only making one of the header lines a
brighter color it gives the blocks a bit more emphasis without becoming too busy.